requirements.txt
openai>=1.0
python-dotenv>=1.0
pymupdf>=1.24
pypdf>=4.0
python-docx>=1.0
